requirements.txt
tensorflow-gpu==2.2.0
numpy
pandas
scikit-learn
seaborn
edit-distance